Update mincho.otf
This rewrites the `ascent` and `descent` values in the `hhea` table of
the font, so that the height of the surface returned by `TTF_Render*`
will be the same as the `ptsize` passed to `TTF_OpenFont`. (MTLc3m.ttf
already have this property.)

This fixes the text rendering issue in Tsumamigui2 and Yoru ga Kuru.